address Gengliang's comments
